Advance dependencies

+++ cmd/main.go
@@ -195,8 +195,15 @@ func initFlags(fs *pflag.FlagSet) {
 }
 
 // fluxCRDHandler restarts process if a Flux CRD is updated
-func fluxCRDHandler(gvk *schema.GroupVersionKind) {
+func fluxCRDHandler(gvk *schema.GroupVersionKind, action crd.ChangeType) {
+	if action == crd.Modify {
+		return
+	}
+
 	if gvk.Group == sourcev1.GroupVersion.Group {
+		setupLog.V(logsettings.LogInfo).Info("Initiating graceful restart due to Flux CRD update",
+			"GVK", gvk.String(), "Action", string(action))
+
 		if killErr := syscall.Kill(syscall.Getpid(), syscall.SIGTERM); killErr != nil {
 			panic("kill -TERM failed")
 		}
